+++ CHANGELOG.md
@@ -3,18 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.5](https://github.com/chrisrzhou/react-wordcloud/compare/v1.0.4...v1.0.5) (2019-03-16)
+
+### New:
+
+- Added `FAQ` page explaining common 'bugs', and updated `Options` page with more examples
+- Handled recursive attempts to layout 'bad' clouds. Provide a console warning when max attempts have been made to layout 'bad' clouds.
+- Changed default `minSize` and `options.fontSizes` value to make things less buggy.
+
+### Bug fixes:
+
+- Fixed a bug where `rotationAngles` was mutated.
+
 ## [1.0.4](https://github.com/chrisrzhou/react-wordcloud/compare/v1.0.3...v1.0.4) (2019-03-16)
 
 Improve and simplify React hooks code after detailed understanding of: https://overreacted.io/a-complete-guide-to-useeffect/
 
-Bug fixes:
+### Bug fixes:
 
 - Handle words that don't fit in the boundary of the SVG by applying a font-size scale factor
 - Handle large number of words
 
 ## [1.0.3](https://github.com/chrisrzhou/react-wordcloud/compare/v1.0.2...v1.0.3) (2019-03-14)
 
-Bug fixes:
+### Bug fixes:
 
 - https://github.com/chrisrzhou/react-wordcloud/issues/5
 - https://github.com/chrisrzhou/react-wordcloud/issues/11